Updated DTO and database columns

+++ src/main/java/com/example/project/dto/EmployeeDTO.java
@@ -6,10 +6,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class EmployeeDTO {
-    private Long id;
+    private int id;
     private String name;
     private int age;
-    private float salary;
+    private int salary;
     private String city;
     private String designation;
